requirements.txt
Flask==3.0.0
psycopg2_binary==2.9.9
python-decouple==3.8
